+++ api/src/db/migrations/20190522174023_unique_tasker_tm_ids.js
@@ -1,0 +1,20 @@
+
+exports.up = async function (knex) {
+  try {
+    await knex.schema.alterTable('campaigns', t => {
+      t.unique(['tasker_id', 'tm_id'])
+    })
+  } catch (e) {
+    console.error(e)
+  }
+}
+
+exports.down = async function (knex) {
+  try {
+    await knex.schema.alterTable('campaigns', t => {
+      t.dropUnique(['tasker_id', 'tm_id'])
+    })
+  } catch (e) {
+    console.error(e)
+  }
+}