+++ api/tests/api/campaigns.test.js
@@ -3,8 +3,9 @@ const request = require('supertest')
 const db = require('../../src/db/connection')
 let app = require('../../src/index')
 const path = require('path')
-const { prop } = require('ramda')
+const { prop, sort, reverse } = require('ramda')
 const { isBefore, isAfter } = require('date-fns')
+const { alphabeticalDiff } = require('../../../lib/utils/sort')
 
 const dbDirectory = path.join(__dirname, '..', '..', 'src', 'db')
 const migrationsDirectory = path.join(dbDirectory, 'migrations')
@@ -107,4 +108,28 @@ test('Get campaigns sorted by least recently updated', async t => {
     t.truthy(isBefore(toCompare, date))
     toCompare = date
   })
+})
+
+test('Get campaigns sorted alphabetically A to Z', async t => {
+  const response = await request(app)
+    .get('/scoreboard/api/campaigns?sortType=Alphabetical A-Z')
+    .expect(200)
+
+  const records = response.body.records
+  const names = records.map(prop('name'))
+
+  const sorted = sort(alphabeticalDiff, names)
+  t.deepEqual(sorted, names)
+})
+
+test('Get campaigns sorted alphabetically Z to A', async t => {
+  const response = await request(app)
+    .get('/scoreboard/api/campaigns?sortType=Alphabetical Z-A')
+    .expect(200)
+
+  const records = response.body.records
+  const names = records.map(prop('name'))
+
+  const sorted = sort(alphabeticalDiff, names)
+  t.deepEqual(reverse(sorted), names)
 })