+++ api/tests/services/tm3.test.js
@@ -2,18 +2,36 @@
  * Test the TM service code for different TM adapters
  */
 const { TM } = require('../../src/services/tm')
+const db = require('../../src/db/connection')
 const test = require('ava')
 const path = require('path')
 const http = require('http')
 const yakbak = require('yakbak')
-const { has } = require('ramda')
+const { has, equals, all } = require('ramda')
 
 const tm3proxy = http.createServer(yakbak('https://tasks.openstreetmap.us', {
   dirname: path.join(__dirname, '..', 'tapes')
 }))
 
+const dbDirectory = path.join(__dirname, '..', '..', 'src', 'db')
+const migrationsDirectory = path.join(dbDirectory, 'migrations')
+const seedsDirectory = path.join(dbDirectory, 'seeds', 'test')
+
+test.before(async t => {
+  await db.migrate.latest({ directory: migrationsDirectory })
+  await db.seed.run({ directory: seedsDirectory })
+})
+
+test.after.always(async t => {
+  await db.destroy()
+})
+
 test.beforeEach.cb(t => {
   tm3proxy.listen(4848, t.end)
+})
+
+test.afterEach.cb(t => {
+  tm3proxy.close(t.end)
 })
 
 test.serial('Test TM3', async t => {
@@ -59,7 +77,7 @@ test.serial('Test TM3 schema', async t => {
 })
 
 test.serial('Test URL forming', async t => {
-  const tm = new TM(1, 'tm3', 'http://tasks.openstreetmap.us', 'http://localhost:4848')
+  const tm = new TM(1, 'tm3', 'http://tasks.openstreetmap.us', { proxy: 'http://localhost:4848' })
   let projects = await tm.getProjects() // Should get from the proxy
 
   const project = projects.find(p => p.projectId === 77)
@@ -67,4 +85,32 @@ test.serial('Test URL forming', async t => {
 
   const url = tm.getUrlForProject(project.projectId)
   t.is(url, `http://tasks.openstreetmap.us/project/77`)
+})
+
+test.serial('Test extra params', async t => {
+  // Sort response with date
+  const tm = new TM(1, 'tm3', 'http://localhost:4848', {
+    search_params: {
+      'mapperLevel': 'BEGINNER'
+    }
+  })
+
+  let projects = await tm.getProjects()
+  let levels = projects.map(p => p.mapperLevel)
+  t.true(all(equals('BEGINNER'), levels))
+})
+
+test.only('Duplicate campaigns', async t => {
+  // count is from seed
+  const [first] = await db('campaigns').count()
+
+  // Try adding the tasks again
+  const [tm3] = await db('taskers').where('name', 'test tm3')
+  const tm = new TM(tm3.id, 'tm3', 'http://tasks.openstreetmap.us', { proxy: 'http://localhost:4848' })
+  let projects = await tm.getProjects() // Should get from the proxy
+  let dbObjects = await tm.toDBObjects(projects)
+  await tm.updateDB(db, dbObjects)
+
+  const [second] = await db('campaigns').count()
+  t.is(first.count, second.count)
 })