+++ components/admin/AdminCampaignsSearch.js
@@ -83,12 +83,10 @@ class CampaignSearch extends Component {
   }
 
   render () {
-    if (!this.props.campaigns) return <div />
-
     const { selectedCampaigns } = this.props
-    const { page } = this.props.campaigns
-    const { records: { total, records } } = this.props.campaigns
-    if (!records) return <div />
+    const { page } = this.props.campaignSearchResults
+    const { records: { total, records } } = this.props.campaignSearchResults
+    if (!records || !selectedCampaigns) return <div />
 
     let assignments = selectedCampaigns.map(record =>
       <Assignment record={record}
@@ -164,4 +162,4 @@ class CampaignSearch extends Component {
   }
 }
 
-export default connect(['campaigns'], actions)(CampaignSearch)
+export default connect(['campaigns', 'campaignSearchResults'], actions)(CampaignSearch)