+++ pages/user.js
@@ -1,6 +1,5 @@
 import React, { Component } from 'react'
 import DashboardHeader from '../components/dashboard/DashboardHeader'
-import getSumEdits from '../lib/utils/sum_edits'
 import ScoreboardPanel from '../components/ScoreboardPanel'
 import EditBreakdownChart from '../components/charts/EditBreakdownChart'
 import CampaignsChart from '../components/charts/CampaignsChart'
@@ -45,7 +44,6 @@ export class User extends Component {
     } = this.props.user
     const { extent_uri, uid } = records
     if (!records) return <div />
-    const editCount = getSumEdits(records)
     const badgeCount = Object.keys(badges.earnedBadges).length
     const campaignCount = records.hashtags.length
     const { name, hashtags, edit_times } = records
@@ -85,38 +83,40 @@ export class User extends Component {
           facets={[
             { label: 'Campaigns', value: formatDecimal(campaignCount) },
             { label: 'Badges', value: formatDecimal(badgeCount) },
-            { label: 'Edits', value: formatDecimal(editCount) }
+            { label: 'Edits', value: formatDecimal(records.edit_sum) },
+            { label: 'Changesets', value: formatDecimal(records.edit_count) }
           ]}
         />
         <div className='row'>
           <DashboardBlurb {...records} username={name} />
-          <CSVLink
-            className='link--large'
-            style={{ float: 'right', marginBottom: '1rem' }}
-            data={[
-              {
-                badgeCount,
-                editCount,
-                campaignCount,
-                records
-              }
-            ]}
-            headers={[
-              { label: 'Name', key: 'records.name' },
-              { label: 'Campaigns', key: 'campaignCount' },
-              { label: 'Badges', key: 'badgeCount' },
-              { label: 'Countries', key: 'records.country_list.length' },
-              { label: 'Roads (Km)', key: 'records.km_roads_add' },
-              { label: 'Buildings', key: 'records.buildings_add' },
-              { label: 'Points of Interest', key: 'records.poi_add' },
-              { label: 'Coastlines (Km)', key: 'records.km_coastlines_add' },
-              { label: 'Waterways (Km)', key: 'records.km_waterways_add' },
-              { label: 'Total Edits', key: 'editCount' }
-            ]}
-            filename={`${name}_ScoreboardData.csv`}
-          >
-            Export User Data (CSV)
-          </CSVLink>
+          <div className='widget-33 page-actions'>
+            <CSVLink
+              className='button button--secondary'
+              style={{ float: 'right', marginBottom: '1rem' }}
+              data={[
+                {
+                  badgeCount,
+                  campaignCount,
+                  records
+                }
+              ]}
+              headers={[
+                { label: 'Name', key: 'records.name' },
+                { label: 'Campaigns', key: 'campaignCount' },
+                { label: 'Badges', key: 'badgeCount' },
+                { label: 'Countries', key: 'records.country_list.length' },
+                { label: 'Roads (Km)', key: 'records.km_roads_add' },
+                { label: 'Buildings', key: 'records.buildings_add' },
+                { label: 'Points of Interest', key: 'records.poi_add' },
+                { label: 'Coastlines (Km)', key: 'records.km_coastlines_add' },
+                { label: 'Waterways (Km)', key: 'records.km_waterways_add' },
+                { label: 'Total Edits', key: 'records.edit_sum' }
+              ]}
+              filename={`${name}_ScoreboardData.csv`}
+            >
+              Export User Data (CSV)
+            </CSVLink>
+          </div>
         </div>
         <section>
           <div className='row'>
@@ -161,11 +161,16 @@ const connectedUser = connect(
   ['user'],
   actions
 )(User)
-connectedUser.getInitialProps = function ({ req }) {
-  const { id } = req.params
-  return {
-    id
-  }
-}
 
 export default connectedUser
+
+connectedUser.getInitialProps = function ({ query }) {
+  if (query) {
+    const { id } = query
+    return {
+      id
+    }
+  } else {
+    return {}
+  }
+}